Create manipulating_df_py
Expectations of code:

Create a DataFrame from the table named ksprojects. Write code that allows you to answer the following questions. Create a separate cell for each question.

How many rows of data are in the DataFrame?

What are the names and data types of the columns?

Do any of the columns contain null values?

Find all successful documentary projects and sort them by the amount pledged. Print the top 10 highest pledges.

Create a new column named average_per_backer and set its value to the total amount pledged / number of backers. What happened to the rows with 0 backers? How can this be dealt with?

Drop all rows with 0 backers then repeat the previous exercise.

Create a crosstab to get a count of records for each combination of state and category.