Create descriptive_stats_py
The expectations of the code are to answer the following questions:

How many rows of data does the table contain?

Are there any nulls in any columns?

How many distinct values are there in the mszoning column?

How many distinct values are there in the street column?

Describe the range of the saleprice.

Describe the range of the lotarea.

Is there a statistically significant difference between the prices of houses that have an open porch versus houses that do not have an open porch? How did you handle null values? Hint: use the variable openporchsf. You will need to transform this variable so that it can be used in a t-test.

Is there a significant correlation between the lotarea and saleprice variables? Are they positively or negatively correlated?