Create fundamentals_evaluation_py
This was a cumulative assessment of my python knowledge. 

The point of this exam is for you to demonstrate your mastery of the basics of programming and data analysis in Python.

The expectations of the code:

Create a function that reverses a string. For example, given the string "forward", your function should return "drawrof". Assume all inputs are strings.

Create a SQL query to combine all four tables in the bitcoinhistoricaldata database. Load the query results into a dataframe in Pandas, and do some exploratory data analysis. Explore the distributions of your variables and at least one correlation. 

Try to find something interesting to say about the data, and use visualizations to support your argument.